+++ src/wam2layers/tracking/backtrack.py
@@ -2,8 +2,8 @@ import logging
 from pathlib import Path
 
 import numpy as np
-import pandas as pd
 import xarray as xr
+from xarray.core.coordinates import DatasetCoordinates
 
 from wam2layers.config import Config
 from wam2layers.tracking.core import (
@@ -20,6 +20,7 @@ from wam2layers.tracking.io import (
     write_output,
 )
 from wam2layers.tracking.shared import initialize_tagging_region, initialize_time
+from wam2layers.utils.calendar import cftime_from_timestamp, round_cftime
 from wam2layers.utils.grid import get_boundary, get_grid_info, stagger_x, stagger_y
 from wam2layers.utils.profiling import ProgressTracker
 
@@ -32,7 +33,7 @@ def backtrack(
     S0,
     tagging_mask,
     output,
-    config,
+    config: Config,
 ):
     a, dy, dx = get_grid_info(F)
 
@@ -113,14 +114,14 @@ def backtrack(
     output["gains"] += gains
 
 
-def initialize(config_file):
+def initialize(config_file: Config) -> tuple[Config, xr.Dataset, DatasetCoordinates]:
     """Read config, region, and initial states."""
     logger.info(f"Initializing experiment with config file {config_file}")
 
     config = Config.from_yaml(config_file)
 
     # Initialize outputs as empty fields based on the input coords
-    t = pd.Timestamp(config.tracking_end_date)
+    t = config.tracking_end_date
     grid = load_data(t, config, "states").coords
     output = xr.Dataset(
         {
@@ -151,7 +152,7 @@ def initialize(config_file):
 #############################################################################
 
 
-def run_experiment(config_file):
+def run_experiment(config_file: Config):
     """Run a backtracking experiment from start to finish."""
     config, output, grid = initialize(config_file)
 
@@ -168,7 +169,7 @@ def run_experiment(config_file):
         tagging_region = initialize_tagging_region(bbox, lat, lon)
         tagging_region_stationary = True
 
-    while t0 >= config.tracking_start_date:
+    while t0 >= cftime_from_timestamp(config.tracking_start_date, config.calendar):
         S0 = load_data(t0, config, "states")
         F = load_data(th, config, "fluxes")
         S1 = load_data(t1, config, "states")
@@ -200,7 +201,7 @@ def run_experiment(config_file):
         t0 -= dt
 
         # Daily output
-        is_output_time = t1 == t1.floor(config.output_frequency)
+        is_output_time = t1 == round_cftime(t1, config.output_frequency, "floor")
         is_final_step = t0 < config.tracking_start_date
         if is_output_time or is_final_step:
             progress_tracker.print_progress(t1, output)